smf: fix warning introduced by last commit
Some platforms have size_t smaller than uint64_t. This cast is fine
because the value originally came from a function returning size_t.

+++ src/smf.cpp
@@ -371,7 +371,7 @@ bool SMF_ReadChunk(SMF_Reader& reader, SMF_Data& data)
     }
     else
     {
-        fprintf(stderr, "Unexpected chunk type at %zu\n", chunk_start);
+        fprintf(stderr, "Unexpected chunk type at %zu\n", (size_t)chunk_start);
         return false;
     }
